# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTransactionRequest:
    type: object
...
  dto.ParseNLPRequest:
    properties:
      text:
        type: string
    required:
    - text
    type: object
  dto.ParseNLPResponse:
    properties:
      amount:
        type: number
      categoryId:
        type: string
      categoryName:
        type: string
      confidenceScore:
        type: number
      description:
        type: string
      type:
        description: INCOME or EXPENSE
        type: string
    type: object
...
  dto.TransactionResponse:
    properties:
      amount:
        example: 100.5
        type: number
      categoryId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      categoryName:
        example: Food
        type: string
      createdAt:
        example: "2024-01-15T00:00:00Z"
        type: string
      deletedAt:
        example: "2024-01-20T00:00:00Z"
        type: string
      description:
        example: Grocery shopping
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      realizedPnl:
        example: 50
        type: number
      status:
        example: HOLDING
        type: string
      transactionDate:
        example: "2024-01-15T00:00:00Z"
        type: string
      type:
        example: EXPENSE
        type: string
      updatedAt:
        example: "2024-01-15T00:00:00Z"
        type: string
      userId:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  dto.UpdateTransactionRequest:
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Nexo API documentation
  title: Nexo API
  version: "1.0"
paths:
...
  /transactions:
    get:
      description: List transactions
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TransactionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction (Income or Expense)
      parameters:
      - description: Create transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new transaction
      tags:
      - transactions
  /transactions/{id}:
    delete:
      description: Delete a transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a transaction
      tags:
      - transactions
    get:
      description: Get a transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a transaction by ID
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Update a transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Update transaction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a transaction
      tags:
      - transactions
  /transactions/parse-nlp:
    post:
      consumes:
      - application/json
      description: Parse Vietnamese text input into structured transaction fields
        (amount, category, type, description)
      parameters:
      - description: Parse NLP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ParseNLPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ParseNLPResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Parse natural language transaction input
      tags:
      - transactions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"